requirements.txt
# for development/running
pypubsub

# for testing
nose

# for packaging
pyinstaller
